recv: Move the block processing out of the callback context.
Since the minclient behavior is to dispatch another `ReceiveMessage`
command after the callback returns successfully to try to keep the
server's state in sync (or to fetch further blocks), how long it
takes to process non-SURB blocks leaks information to the provider.

This attempts to avoid that problem by storing blocks received off
the wire in a queue, and doing the block processing outside of the
worker context.