api: Add `IsConnected()`.
Exactly what it says on the tin.  Eventually there will be events for
this as well, but having an explicit query method is also useful.

Note: The connection to the provider being up does not guarantee that
sends will succeed, though there's no way presently to detect the
network state beyond the provider.