api: Add `config.Proxy.EventSink`, and `events/`.
If there is a channel set, mailproxy will now feed events (connection
status, message receiption) to the channel.  There probably will be more
events added in the future, but this should be the bare minimum for now.

Note: If the caller is slow to consume off the event sink, events
can/will get lost, starting from the eldest first (drop-head), because
"not screwing up mailproxy's state" is more important.

Part of #11.