update registration_client
mailproxy would previously trample a configuration file if it was asked
to register a new account. v0.0.3 of registration_client will instead
return an error if any existing key material or configuration is found.

TODO: replace the configuration generation section of
registration_client so that it may append a new account to the
configuration, or build this feature into mailproxy.